refactor(hub): move to @/components
